Se cambia la vista de estado de cupo y se muestra correctamente la vista de cupo aprovado y rechazo definitivo

+++ app/src/main/java/com/pagatodo/yaganaste/data/model/webservice/response/cupo/DataEstadoSolicitud.java
@@ -25,7 +25,7 @@ public class DataEstadoSolicitud  implements Serializable{
     private int IdEstadoCivil;
     private int IdEstatusPaso;
     private int IdPaso;
-    private int LineaCredito;
+    private Float LineaCredito;
     private String NumeroCliente;
     private int NumeroHijos;
     private String Paso;
@@ -115,11 +115,11 @@ public class DataEstadoSolicitud  implements Serializable{
         IdPaso = idPaso;
     }
 
-    public int getLineaCredito() {
+    public Float getLineaCredito() {
         return LineaCredito;
     }
 
-    public void setLineaCredito(int lineaCredito) {
+    public void setLineaCredito(Float lineaCredito) {
         LineaCredito = lineaCredito;
     }
